requirements.txt
absl-py==0.12.0
asgiref==3.2.3
astroid==2.0.4
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.0
bs4==0.0.1
cachetools==4.2.1
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.3
coverage==5.0.3
decorator==4.4.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.37.1
django-cors-headers==2.4.0
django-extensions==2.2.6
django-heroku==0.3.1
django-rest-auth==0.9.3
djangorestframework==3.9.1
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
geographiclib==1.50
geopy==1.21.0
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==19.7.1
h5py==2.10.0
idna==2.7
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.3.1
libpysal==4.2.2
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.9
notebook==6.0.3
numpy==1.19.5
oauthlib==2.1.0
opt-einsum==3.3.0
pandas==1.0.5
pandocfilters==1.4.2
parso==0.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.15.6
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pylint==2.1.1
pyrsistent==0.15.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.5
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.7.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sklearn-crfsuite==0.3.6
soupsieve==2.0.1
sqlparse==0.3.0
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.1.0
typing-extensions==3.7.4.3
urllib3==1.24.2
virtualenv==16.7.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==3.3.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==0.90
zipp==2.1.0